Fixes bug 29 and bug 30.
This separates the ticktick function parser to two
modes: an "in string" and an "out of string" intepretation.

Out of string, things work very much like they always have.

In string, however, content gets fed into the token looking for
the end of string, or an escaped (\) string.  This then gets
passed off to the sanitization routine.

It meant that backslashes had to be moved in the tokenization chain
so as to always deal with them individually.

Additionally, when the json_parse_value gets entered, the old code would
remove all quotations, whether escaped or not.  The new code tries to be
smarter about that so that the expanded variable names will line up
correctly.